Fix: Several small codebase fixes (#413)
* Several codebase fixes

* ci: update coverage badges

* Use injected loggers for lifecycle cleanup

This change removes lazy logger lookups and instead uses injected ILogger instances across the WebView manager and related window features. It also replaces Debug.WriteLine calls in disposal/finalizer paths with structured logging when a logger is available, and resolves logger instances from the window service provider in monitor and external URL handlers. Tests were updated to pass the new logger dependency for the WebView manager.

* Update InfiniFrameWebViewManagerTests.cs

* Add external launcher abstraction and expand tests

This change introduces an abstraction for external process startup and registers it in DI so web-message handlers can be unit tested without launching real browser processes. It also modernizes several array and collection initializations to use collection expressions and adds broader coverage around native interop status, static asset manifest handling, file picker/filter behavior, and web-message dispatch flows.

* Update Win32SetWebView2PathTests.cs

* Update Win32SetWebView2PathTests.cs

* Use ubuntu runner; fix coverage script

Change coverage workflow to run on ubuntu-latest and harden the coverage_report script. Adds explicit encoding="utf-8" to subprocess.run for consistent UTF-8 handling and checks that gh API output is non-empty before calling strip to avoid errors when stdout is None or empty. Files changed: .github/workflows/shared-coverage.yml, scripts/coverage_report.py.

* Move external process launcher to Interop

Refactors external process launching into the InfiniFrame.Interop namespace and adds the concrete launcher implementation to the shared project. The service registration is updated to resolve the interface from the new namespace, and related tests are adjusted to match the new contracts and analyzer cleanups.

* ci: update coverage badges

* Exclude native files from coverage

This change excludes native C/C++/Objective-C source files from Cobertura coverage parsing so they do not inflate coverage totals. It also adds tests covering the new exclusion behavior for native file filtering alongside the existing test package exclusions.

* ci: update coverage badges

* Deduplicate Cobertura line coverage

Avoid inflated line coverage by deduplicating tracked line numbers per class before counting covered and total lines. This keeps method-level duplicates from inflating IL while preserving coverage from any positive-hit line. Added a regression test covering duplicate line numbers across methods in the same class.

* Update sync_github_checks.py

* Probe Playwright CDP endpoint before tests

Add a probing step in BlazorPlaywrightContextBase to verify the Playwright CDP endpoint is reachable after startup. Uses PlaywrightConnectionUtility.CreateCdpConnectionUrl and RemoteDebuggingUtility.TryProbeEndpoint in a 60s loop, logging success or a warning and proceeding so Playwright can retry. Also add the necessary InfiniFrame.Utilities import.

* ci: update coverage badges

---------

+++ src/InfiniFrame.NativeBridge/Managed/Delegates/CppNavigationStartingDelegate.cs
@@ -16,7 +16,7 @@ namespace InfiniFrame.NativeBridge.Delegates;
 /// <param name="isMainFrame">Non-zero if the navigation is in the main frame.</param>
 /// <returns>0 to allow, 1 to cancel.</returns>
 [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-internal delegate byte CppNavigationStartingDelegate(
+internal delegate int CppNavigationStartingDelegate(
     [MarshalAs(UnmanagedType.LPUTF8Str)]
     string url,
     int isUserInitiated,
